backend/money_transfer/requirements.txt
asgiref==3.9.0
Django==5.2.4
sqlparse==0.5.3
tzdata==2025.2
psycopg2-binary==2.9.9
django-environ==0.11.2
django-cors-headers==4.3.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
python-dotenv==1.0.1
pillow==11.3.0
